autonomous_research_agent_/autonomous_research_agent/requirements.txt
langchain
langchain-openai
langchain-anthropic
langchain-community
tavily-python
wikipedia
python-dotenv
